Add BOTS file
This new BOTS file contains the entries from IntelMQ's BOTS file from
IntelMQ commit 3f9eb71aff966c7510d281658c1242eb168a0e3f for the
CERT-Bund bots that are now maintained in this repository. The only
difference is the new module name due to the different directory layout.

In order to use the bots with IntelMQ the contents of this BOTS file
have to be integrated in the obvious way into IntelMQ's BOTS file.