Associate tags, not annotations with emails
Looking at how annotations have been used in practice so far, it turns
out that it suffices to have tags instead of the annotations that we
have now (which have a tag, but may contain more fields). So we limit
annotations for email addresses to just the tags.

The tags are further structured by categories, so that each tag
basically has the form "category:tag". There are two tables now for the
categeories and the tags in the categories and the email_annotation
table has been replaced with email_tag.

To make it easier for contact database bot, there's also a view that
shows the tags as annotations, so that the bots only have to deal with
annotations for now which should make working with the contact data more
consistent in the bots.