Allow tags to be marked as their category's default tag
This commit adds a flag, is_default, to the tag table. All tags where
this flag is true are considered the default tag of the corresponding
category. There's a new partial unique index that makes sure that
there's at most one default tag per category.

In order to make determining the annotations of an email address
simpler, there's now a stored procedure that determines the annotations
for an email address which looks up the tags in the email_tag table and
supplements that with any default tags for categories for which no tag
was specified in email_tag.