Handle abuse-c values without entries in ripe.db.role in ripe_data.py
It can happen that entries in ripe.db.inetnum.gz or ripe.db.inet6num.gz
have abuse-c references that do not occur in the ripe.db.role.gz file.
The only place this actually mattered in the code is in
modify_for_abusec where new organisations are created in some cases and
the name of the new organisations is derived from the role entrey. The
code in ripe_data.py now handles the missing entry now by inventing a
suitable name.

It may be possible for other files to be affected as well sometimes, but
it did actually occur in the two mentioned above.