Discard inetnum/inet6num entries with unknown abuse-c
Entries from ripe.db.inetnum.gz and ripe.db.inet6num.gz which have an
abuse-c attribute that is not present in ripe.db.role.gz are now
discarded when importing the data. The code prints a message about this.

The previous change tried to keep the data even though it would not be
possible to send notifications. The new behaviour of discarding the
entries altogether is seen as the better choice.