Rework the names the email tag tables and columns.
This renames the category table to tag_name and adjusts the column names
accordingly and also the places where these names are used in
intelmq-certbund-contact.

The conceptual model for the email tags used to be tags classified into
categories, now it's the idea that the tags represent "key:value" pairs
where the key is the former category and the value the former tag. The
new model does fit the way the tags associated with the email addresses
are used and the constraints of the tag and tag_name tables (there are
default values, for instance).