Add script to create email tag test cases
As #11 has a number of cases that we want to detect
and possibly handle differently, test cases are needed.

To create them reliably and without we use a script that creates
sql commands. An alternative is have been to go through the
fody-backend, but which would require one more component to run
for building the test cases.