ripe-import: Add --before-commit-command option
The RIPE importer has a new command line option,
--before-commit-command, whose argument is an SQL statement that is
executed by the importer before the transaction that updates the RIPE
data is committed. This command can be used to e.g. cleanup data that
depends on the RIPE data, but which cannot be automatically updated with
'ON DELETE CASCADE' for foreign keys or similar mechanisms.