Don't look up pytest on $PATH
If running with a venv active that does not have pytest installed, but
you have it installed outside the venv, looking up "pytest" on $PATH
will find nothing inside the venv and then successfully find it outside,
leading to your global Python installation being used rather than the
one from the venv. With this change, we force the use of the "pytest"
module that belongs to whatever "python3" currently is (this will be
replaced inside an active venv, which is what we want), which will cause
a clear error message if pytest is not installed for it, rather than a
confusing one (e.g., "no module intelmq" even though the venv has it,
because the global Python installation doesn't).